Add the code needed to run the rock paper scissors
First, I created a math function that returns a
random number from 0 to 2. In other word, 3 random
values. I assigned that function to a variable.
After, I assigned rock, paper and scissors keyword
to each number in an if else statement.Then, I
created a variable that captures the user input of
what they choose to play in the game. Afterwise, I
Created another if else statement to return the
variable that randomly generates one of the game
responses when the user inputs their play.